requirement.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2026.1.4
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
frozenlist==1.8.0
idna==3.11
kiwisolver==1.4.9
magic-filter==1.0.12
matplotlib==3.10.8
multidict==6.7.0
numpy==2.4.0
packaging==25.0
pillow==12.1.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
